requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
Pillow==5.4.1
pyTelegramBotAPI==3.6.6
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
